Added check for es setting before configuring es

+++ test/conftest.py
@@ -1,14 +1,16 @@
 def pytest_configure(config):
     from rockpack.mainsite import app, init_app
-    from rockpack.mainsite.core.es import mappings, helpers
 
-    mappings.CHANNEL_INDEX = 'test_channels'
-    mappings.VIDEO_INDEX = 'test_videos'
-    mappings.USER_INDEX = 'test_users'
+    if app.config.get('ELASTICSEARCH_URL'):
+        from rockpack.mainsite.core.es import mappings, helpers
 
-    i = helpers.Indexing()
-    i.create_all_indexes(rebuild=True)
-    i.create_all_mappings()
+        mappings.CHANNEL_INDEX = 'test_channels'
+        mappings.VIDEO_INDEX = 'test_videos'
+        mappings.USER_INDEX = 'test_users'
+
+        i = helpers.Indexing()
+        i.create_all_indexes(rebuild=True)
+        i.create_all_mappings()
 
     app.config['DATABASE_URL'] = 'sqlite://'
 
@@ -26,8 +28,10 @@ def pytest_configure(config):
 
 
 def pytest_unconfigure(config):
-    from rockpack.mainsite.core.es import helpers
-    i = helpers.Indexing()
-    i.delete_index('channel')
-    i.delete_index('video')
-    i.delete_index('user')
+    from rockpack.mainsite import app
+    if app.config.get('ELASTICSEARCH_URL'):
+        from rockpack.mainsite.core.es import helpers
+        i = helpers.Indexing()
+        i.delete_index('channel')
+        i.delete_index('video')
+        i.delete_index('user')
